# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a get role server.",
        "title": "GetRole API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "stoola08@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/role": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get all roles",
                "operationId": "getRoles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/role/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "Get one role based on ID",
                "operationId": "getRole",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": ""
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}